requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
argbind==0.3.9
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
base64io==1.0.3
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
distro==1.9.0
docstring_parser==0.16
einops==0.8.0
elevenlabs==1.7.0
exceptiongroup==1.2.2
executing==2.1.0
ffmpy==0.4.0
filelock==3.15.4
fire==0.6.0
flatten-dict==0.4.2
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
groq==0.11.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
kiwisolver==1.4.5
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37
langchain-groq==0.1.9
langchain-text-splitters==0.2.2
langsmith==0.1.108
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
parler_tts @ git+https://github.com/huggingface/parler-tts.git@8e465f1b5fcd223478e07175cb40494d19ffbe17
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
playsound==1.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==3.19.6
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py3-tts==3.5
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pyparsing==3.1.4
pystoi==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.91
PyYAML==6.0.2
randomname==0.2.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==2.0.32
stack-data==0.6.3
sympy==1.13.2
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torch-stoi==0.2.1
torchaudio==2.4.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.43.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
websockets==13.0.1
Werkzeug==3.0.4
yarl==1.9.7
zipp==3.20.1
